Moved extensions to separate files.
